# docs/swagger.yaml
definitions:
  account.AccountResponse:
    properties:
      balance:
        type: number
      cpf:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  account.BalanceResponse:
    properties:
      balance:
        type: number
    type: object
  login.LoginResponse:
    properties:
      token:
        type: string
    type: object
  tranfer.TransferResponse:
    properties:
      account_destination_id:
        type: string
      account_origin_id:
        type: string
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
    type: object
info:
  contact: {}
paths:
  /account:
    get:
      consumes:
      - application/json
      description: get list of accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.AccountResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Show accounts
      tags:
      - accounts
  /account/:
    post:
      consumes:
      - application/json
      description: Create account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.AccountResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create account
      tags:
      - accounts
  /account/{account_id}/balance:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.BalanceResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Show balance from account
      tags:
      - accounts
  /login:
    post:
      consumes:
      - application/json
      description: User Login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/login.LoginResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: User Login
      tags:
      - login
  /transfer/:
    get:
      consumes:
      - application/json
      description: Show all transfer from user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tranfer.TransferResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Show all transfer from user
      tags:
      - transfer
    post:
      consumes:
      - application/json
      description: Make a transfer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tranfer.TransferResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Make a transfer
      tags:
      - transfer
swagger: "2.0"